+++ src/typing_ex/view/page.clj
@@ -6,7 +6,7 @@
    [ring.util.anti-forgery :refer [anti-forgery-field]]
    [taoensso.timbre :as timbre]))
 
-(def ^:private version "1.3.4")
+(def ^:private version "1.3.8")
 
 (defn page [& contents]
   [::response/ok
@@ -48,7 +48,13 @@
      "成績に影響しない欠席ひとつに神経質になるより、"
      "しっかりタイピング平常点稼いだ方が建設的。"]]))
 
-(defn scores-page [ret user days]
+(defn- count-ex-days [days login]
+  (->> days
+       (filter #(= (:login %) login))
+       count))
+
+(defn scores-page [ret user days ex-days]
+  ;;(timbre/debug ex-days)
   (page
    [:h2 "Typing: Scores (last " days " days)"]
    [:p
@@ -56,7 +62,10 @@
     " "
     [:a {:href "/logout" :class "btn btn-warning btn-sm"} "logout"]
     " "
-    [:a {:href "/trials" :class "btn btn-danger btn-sm"} "trials"]
+    [:a {:href "/trials" :class "btn btn-danger btn-sm"} "last40"]
+    " "
+    [:a {:href "/daily" :class "btn btn-danger btn-sm"} "todays"]
+
     [:span {:class "mmm"} " "]
     [:a {:href "http://qa.melt.kyutech.ac.jp/"
          :class "btn btn-info btn-sm"}
@@ -70,13 +79,14 @@
          :class "btn btn-info btn-sm"}
      "L22"]]
 
-   [:p "直近の " days " 日間に練習したユーザのリスト。スコア順。
-名前をクリックすると全データ表示。"]
+   [:p "直近 " days " 日間のスコア順リスト。カッコは通算練習日数。"]
    (into [:ol
           (for [{:keys [max login]} ret]
             [:li
              max
-             " "
+             "("
+             (count-ex-days ex-days login)
+             ") "
              [:a {:href (str "/record/" login)
                   :class (cond
                            (= login user) "yes"
@@ -102,7 +112,7 @@
 
 ;; not good
 (defn- ss [s]
-  (subs (str s) 0 19))
+  (subs (str s) 0 16))
 
 ;; 平均を求めるのに、DB 引かなくても ret から求めればいい。
 ;; ret は lazySeq
@@ -115,19 +125,27 @@
      [:div (plot 300 150 positives)]
      [:br]
      [:ul
-      [:li "Exercises &nbsp;" (count positives)]
-      [:li "Last Exercise &nbsp;" (ss (str (:timestamp (last ret))))]
-      [:li "Last 10 Average &nbsp;" avg]
-      [:li "Max &nbsp;" (apply max (map :pt positives))]]
-     [:p [:a {:href "/" :class "btn btn-primary btn-sm"} "Go!"]]
-     #_(into
-        [:ol {:reversed "reversed"}]
-        (for [{:keys [pt timestamp]} (reverse ret)]
-          [:li pt ", " (ss timestamp)])))))
+      [:li "Max " (apply max (map :pt positives))]
+      [:li "Average (last 10) " avg]
+      [:li "Exercises " (count positives)]
+      [:li "Last Exercise " (ss (str (:timestamp (last ret))))]]
+     [:p [:a {:href "/" :class "btn btn-primary btn-sm"} "Go!"]])))
 
 (defn active-users-page [ret]
   (page
-   [:h2 "Typing: last trials"]
+   [:h2 "Typing: Last 40 trials"]
+   [:p "最近の tp ユーザ 40 名。連続するセッションを１つとするが、
+        セッションの間に別ユーザが割り込むと別セッションとカウント。
+        改良するか？"]
    (into [:ol]
          (for [[u & _] ret]
-           [:li (:login u) " " (ss (:timestamp u))]))))
+           [:li (ss (:timestamp u)) " " (:login u)]))))
+
+(defn todays-act-page [ret]
+  ;;(timbre/debug ret)
+  (page
+   [:h2 "Typing: todays"]
+   [:p "本日の tp ユーザ。重複を省いて最終利用時間で並べ替え。"]
+   (into [:ol]
+         (for [r ret]
+           [:li (ss (:timestamp r)) " " (:login r)]))))